ChatGPT generated versions
They need some work but they were actually good enough I think to provide some
ideas at least to steal so adding them anyway.

If you want to play more with it the prompt I used was:

```
Express Journey's "Don't Stop Believin" as Ruby code. Minimize putting the
lyrics in strings but instead make it so you "hear" the song as you read the
code itself. It does not need to to output the lyrics. The code itself follows
the lyrics.
```

With the polyglot theme of the conference I then did a follow-up to see if it
could modify that to be a bunch of languages (I think the final intention of
the shirt). Gave it the follow-up prompt:

```
Update this so each lyric is in a different programming language and note the
language in a comment.
```

This results in the `dontstopbelieving-gpt.wtf` file. This one needs even more
work as the last example is obviously not assembly, PHP should have
semi-colon's, the comment syntax is a bit wonky but still not bad.

If I get time later I may to a follow-up PR that cleans up the attempts with
my own adjustments. Sort of an eric/gpt collaboration. But going ahead and
sending up the raw GPT versions for now in case time is never made.